refactor: routes 담당 컴포넌트 생성
app 컴포넌트가 너무 복잡하여 routes 만 담당하는 컴포넌트 작성합니다.
